+++ docs/NEWS.rst
@@ -13,6 +13,9 @@ Changes in 5.59 (released 06/30/2020)
 
 *	:mod:`ll.orasql` now supports :mod:`cx_Oracle` 8.0.
 
+*	:meth:`ll.orasql.Connection.objects` now outputs :class:`TableComment`
+	objects too.
+
 
 Changes in 5.58 (released 06/12/2020)
 -------------------------------------